refactor: devide custom hooks files
