+++ src/hooks/useLockBodyScroll.ts
@@ -1,0 +1,11 @@
+import { useLayoutEffect } from 'react'
+
+export const useLockBodyScroll = () => {
+  useLayoutEffect((): (() => void) => {
+    const originalStyle = window.getComputedStyle(document.body).overflow
+
+    document.body.style.overflow = 'hidden'
+
+    return () => (document.body.style.overflow = originalStyle)
+  }, [])
+}
